+++ inst/examples/ex-jackknife.R
@@ -1,0 +1,4 @@
+## Jackknife
+x <- rnorm(20)
+jack <- jackknife(x, do = mean) # Sample mean
+summary(jack)